refactor(fmt): centralize path helpers
