+++ packages/rstack/src/fmt/pathHelpers.ts
@@ -2,6 +2,9 @@ import path from 'node:path';
 
 type RelativePathResolver = (filePath: string) => string;
 
+const toPosixPath: (filePath: string) => string =
+  path.sep === '\\' ? (filePath) => filePath.replaceAll('\\', '/') : (filePath) => filePath;
+
 const createRelativePathResolver = (rootPath: string): RelativePathResolver => {
   const rootPrefix = rootPath.endsWith(path.sep) ? rootPath : `${rootPath}${path.sep}`;
 
@@ -13,5 +16,5 @@ const createRelativePathResolver = (rootPath: string): RelativePathResolver => {
         : path.relative(rootPath, filePath);
 };
 
-export { createRelativePathResolver };
+export { createRelativePathResolver, toPosixPath };
 export type { RelativePathResolver };
